requirements.txt
# ── ORM / Database ─────────────────────────────────────────────────────────────
sqlmodel>=0.0.18
psycopg2-binary>=2.9

# ── Validation / Settings ──────────────────────────────────────────────────────
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0

# ── Migrations ────────────────────────────────────────────────────────────────
alembic>=1.13

# ── Excel / Data Layer ────────────────────────────────────────────────────────
openpyxl>=3.1
pandas>=2.0

# ── AI / Motor Cognitivo ──────────────────────────────────────────────────────
google-genai>=1.0

# ── Telegram Bot ──────────────────────────────────────────────────────────────
python-telegram-bot>=21.0

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit>=1.32
plotly>=5.18
